requirements.txt
altair==4.2.2
attrs==23.1.0
blinker==1.6.2
branca==0.6.0
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
deepdiff==6.3.0
entrypoints==0.4
et-xmlfile==1.1.0
folium==0.14.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
networkx==3.1
numpy==1.24.3
openpyxl==3.1.2
ordered-set==4.1.0
packaging==23.1
pandapower==2.12.1
pandas==2.0.1
Pillow==9.5.0
protobuf==3.20.3
pyarrow==12.0.0
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.30.0
rich==13.3.5
scipy==1.10.1
six==1.16.0
smmap==5.0.0
streamlit==1.22.0
streamlit-folium==0.11.1
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.2
validators==0.20.0
zipp==3.15.0
